+++ src/matterbridgePlatform.ts
@@ -1,10 +1,10 @@
 /**
- * This file contains the class MatterbridgeAccessoryPlatform.
+ * This file contains the class MatterbridgePlatform.
  *
  * @file matterbridgePlatform.ts
  * @author Luca Liguori
  * @date 2024-03-21
- * @version 1.2.0
+ * @version 1.2.1
  *
  * Copyright 2024, 2025, 2026 Luca Liguori.
  *
@@ -74,7 +74,7 @@ export class MatterbridgePlatform {
   storage: NodeStorageManager;
   context?: NodeStorage;
 
-  // Device and entity selection
+  // Device and entity select in the plugin config UI
   readonly selectDevice = new Map<string, { serial: string; name: string; configUrl?: string; icon?: string; entities?: { name: string; description: string; icon?: string }[] }>();
   readonly selectEntity = new Map<string, { name: string; description: string; icon?: string }>();
 
@@ -87,9 +87,6 @@ export class MatterbridgePlatform {
   // Registered devices
   private readonly _registeredEndpoints = new Map<string, MatterbridgeEndpoint>(); // uniqueId, MatterbridgeEndpoint
   private readonly _registeredEndpointsByName = new Map<string, MatterbridgeEndpoint>(); // deviceName, MatterbridgeEndpoint
-
-  // Stored devices
-  private readonly _storedDevices = new Map<string, { pluginName: string; serial: string; name: string; configUrl?: string }>(); // serial, { serial, name }
 
   /**
    * Creates an instance of the base MatterbridgePlatform.
@@ -149,7 +146,7 @@ export class MatterbridgePlatform {
   /**
    * This method must be overridden in the extended class.
    * It is called when the platform is started.
-   * Use this method to create the MatterbridgeDevice and call this.registerDevice().
+   * Use this method to create the MatterbridgeEndpoints and call this.registerDevice().
    * @param {string} [reason] - The reason for starting.
    * @throws {Error} - Throws an error if the method is not overridden.
    */
@@ -159,9 +156,9 @@ export class MatterbridgePlatform {
   }
 
   /**
-   * This method can be overridden in the extended class. Call super.onConfigure() to run checkEndpointNumbers().
+   * This method can be overridden in the extended class. In this case always call super.onConfigure() to save the select and run checkEndpointNumbers().
    * It is called after the platform has started.
-   * Use this method to perform any configuration of your devices and to override the value of the attributes that are stored in the
+   * Use this method to perform any configuration of your devices and to override the value of the attributes that are persistent and stored in the
    * matter storage (i.e. the onOff attribute of the OnOff cluster).
    */
   async onConfigure() {
@@ -177,7 +174,7 @@ export class MatterbridgePlatform {
   /**
    * This method can be overridden in the extended class. In this case always call super.onShutdown() to save the selects, run checkEndpointNumbers() and cleanup memory.
    * It is called when the platform is shutting down.
-   * Use this method to clean up any resources.
+   * Use this method to clean up any resources you used in the constructor or onStart.
    * @param {string} [reason] - The reason for shutting down.
    */
   async onShutdown(reason?: string) {
@@ -194,7 +191,6 @@ export class MatterbridgePlatform {
     this.selectEntity.clear();
     this._registeredEndpoints.clear();
     this._registeredEndpointsByName.clear();
-    this._storedDevices.clear();
 
     // Close the storage
     await this.context?.close();
@@ -203,7 +199,7 @@ export class MatterbridgePlatform {
   }
 
   /**
-   * Sets the logger level and logs a debug message indicating that the plugin doesn't override this method.
+   * Called when the logger level is changed.
    * @param {LogLevel} logLevel The new logger level.
    */
   async onChangeLoggerLevel(logLevel: LogLevel) {
@@ -247,7 +243,15 @@ export class MatterbridgePlatform {
   }
 
   /**
-   * Check if a device with this name is already registered in the platform.
+   * Retrieves the devices registered with the platform.
+   * @returns {MatterbridgeEndpoint[]} The registered devices.
+   */
+  getDevices(): MatterbridgeEndpoint[] {
+    return Array.from(this._registeredEndpoints.values());
+  }
+
+  /**
+   * Checks if a device with this name is already registered in the platform.
    * @param {string} deviceName - The device name to check.
    * @returns {boolean} True if the device is already registered, false otherwise.
    */
@@ -317,7 +321,7 @@ export class MatterbridgePlatform {
   }
 
   /**
-   * Clears the select device and entity maps.
+   * Clears all the select device and entity maps.
    *
    * @returns {void}
    */
@@ -446,13 +450,6 @@ export class MatterbridgePlatform {
   }
 
   /**
-   * @deprecated This method is deprecated and will be removed in future versions. Use validateDevice instead.
-   */
-  validateDeviceWhiteBlackList(device: string | string[], log = true): boolean {
-    return this.validateDevice(device, log);
-  }
-
-  /**
    * Validates if a device is allowed based on the whitelist and blacklist configurations.
    * The blacklist has priority over the whitelist.
    *
@@ -482,13 +479,6 @@ export class MatterbridgePlatform {
     }
     if (log) this.log.info(`Skipping device ${CYAN}${device.join(', ')}${nf} because not in whitelist`);
     return false;
-  }
-
-  /**
-   * @deprecated This method is deprecated and will be removed in future versions. Use validateEntity instead.
-   */
-  validateEntityBlackList(device: string, entity: string, log = true): boolean {
-    return this.validateEntity(device, entity, log);
   }
 
   /**
@@ -525,7 +515,7 @@ export class MatterbridgePlatform {
    *
    * @returns {Promise<number>} The size of the updated endpoint map, or -1 if storage is not available.
    */
-  async checkEndpointNumbers() {
+  async checkEndpointNumbers(): Promise<number> {
     if (!this.storage) return -1;
     this.log.debug('Checking endpoint numbers...');
     const context = await this.storage.createStorage('endpointNumbers');